requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
argparse==1.2.1
gevent==1.0.1
greenlet==0.4.2
gunicorn==19.0.0
itsdangerous==0.24
pyflare==1.0.2
redis==2.10.1
requests==2.3.0
rq==0.4.6
wsgiref==0.1.2
